feat(homepage): review comments: - only one section opened at a time - background of xtmhub logo should be the same as in the menu - change icon of collapse button - new filigran icon/ui versions - use hook to build the menu
