+++ apps/frontend/package.json
@@ -25,8 +25,8 @@
     "i18n:check": "i18n-check --locales ./messages --source en"
   },
   "dependencies": {
-    "@filigran/icon": "0.23.0",
-    "@filigran/ui": "1.2.0",
+    "@filigran/icon": "0.24.0",
+    "@filigran/ui": "1.3.0",
     "@hookform/resolvers": "5.4.0",
     "@svgr/webpack": "8.1.0",
     "@tailwindcss/postcss": "4.3.1",
