+++ apps/frontend/src/components/menu/CollapseMenuButton.tsx
@@ -1,16 +1,18 @@
 import { cn } from '@/lib/utils';
-import { KeyboardArrowLeftIcon } from '@filigran/icon';
+import { KeyboardArrowLeftIcon, LeftPanelCloseIcon } from '@filigran/icon';
 import { Button } from '@filigran/ui';
 import { useTranslations } from 'next-intl';
 
 interface CollapseMenuButtonProps {
   open: boolean;
   handleOpenMenu: () => void;
+  isHomepageV2?: boolean;
 }
 
 export const CollapseMenuButton = ({
   open,
   handleOpenMenu,
+  isHomepageV2 = false,
 }: CollapseMenuButtonProps) => {
   const t = useTranslations();
   return (
@@ -21,12 +23,21 @@ export const CollapseMenuButton = ({
         className="h-9 px-m w-full justify-start rounded-none text-foreground"
         onClick={handleOpenMenu}>
         <span className="flex w-8 shrink-0 justify-center">
-          <KeyboardArrowLeftIcon
-            className={cn(
-              'h-4 w-4 p-1 duration-300 ease-in-out',
-              open ? 'rotate-0' : 'rotate-180'
-            )}
-          />
+          {isHomepageV2 ? (
+            <LeftPanelCloseIcon
+              className={cn(
+                'h-6 w-6 p-1 duration-300 ease-in-out',
+                open ? 'rotate-0' : 'rotate-180'
+              )}
+            />
+          ) : (
+            <KeyboardArrowLeftIcon
+              className={cn(
+                'h-4 w-4 p-1 duration-300 ease-in-out',
+                open ? 'rotate-0' : 'rotate-180'
+              )}
+            />
+          )}
         </span>
         <span className={cn('normal-case', open ? 'ml-2' : 'sr-only')}>
           {t('App.Collapse')}
