+++ apps/frontend/src/components/menu/MenuLogo.tsx
@@ -1,11 +1,20 @@
 import { DisplayLogo } from '@/components/ui/DisplayLogo';
+import { cn } from '@/lib/utils';
 import { useLocale } from 'next-intl';
 import Link from 'next/link';
 
-export const MenuLogo = () => {
+interface MenuLogoProps {
+  withDarkBackground?: boolean;
+}
+
+export const MenuLogo = ({ withDarkBackground = true }: MenuLogoProps) => {
   const locale = useLocale();
   return (
-    <div className="flex z-10 shrink-0 sticky top-0 h-16 border-border-light border-b bg-page-background dark:bg-background items-center px-m">
+    <div
+      className={cn(
+        'flex z-10 shrink-0 sticky top-0 h-16 border-border-light border-b bg-page-background items-center px-m',
+        withDarkBackground && 'dark:bg-background'
+      )}>
       <Link
         href={`/${locale}`}
         aria-label="XTM Hub by Filigran"
